Move iree/base/ref_ptr.h to iree/hal/vulkan/util/. The last user of it; #4369 is tracking replacing DynamicSymbols, the handle types could be ported to unique_ptr semantics, and then the rest is the timeline semaphores fallback.
