+++ iree/hal/CMakeLists.txt
@@ -63,10 +63,10 @@ iree_cc_library(
     half::includes
     iree::base::api
     iree::base::core_headers
-    iree::base::ref_ptr
     iree::base::synchronization
     iree::base::threading
     iree::base::tracing
+    iree::hal::vulkan::util::ref_ptr
   PUBLIC
 )
