+++ iree/hal/vulkan/BUILD
@@ -94,12 +94,12 @@ cc_library(
         "//iree/base:core_headers",
         "//iree/base:flatcc",
         "//iree/base:logging",
-        "//iree/base:ref_ptr",
         "//iree/base:status",
         "//iree/base:synchronization",
         "//iree/base:tracing",
         "//iree/hal:api",
         "//iree/hal/vulkan/util:intrusive_list",
+        "//iree/hal/vulkan/util:ref_ptr",
         "//iree/schemas:spirv_executable_def_c_fbs",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/container:flat_hash_map",
@@ -126,9 +126,9 @@ cc_library(
     deps = [
         "//iree/base:core_headers",
         "//iree/base:dynamic_library",
-        "//iree/base:ref_ptr",
         "//iree/base:status",
         "//iree/base:tracing",
+        "//iree/hal/vulkan/util:ref_ptr",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/memory",
         "@com_google_absl//absl/strings",
