+++ iree/hal/vulkan/CMakeLists.txt
@@ -87,12 +87,12 @@ iree_cc_library(
     iree::base::core_headers
     iree::base::flatcc
     iree::base::logging
-    iree::base::ref_ptr
     iree::base::status
     iree::base::synchronization
     iree::base::tracing
     iree::hal::api
     iree::hal::vulkan::util::intrusive_list
+    iree::hal::vulkan::util::ref_ptr
     iree::schemas::spirv_executable_def_c_fbs
     vulkan_memory_allocator
   PUBLIC
@@ -115,9 +115,9 @@ iree_cc_library(
     absl::strings
     iree::base::core_headers
     iree::base::dynamic_library
-    iree::base::ref_ptr
     iree::base::status
     iree::base::tracing
+    iree::hal::vulkan::util::ref_ptr
   PUBLIC
 )
