+++ iree/hal/vulkan/handle_util.h
@@ -28,12 +28,12 @@
 #include "iree/hal/vulkan/vulkan_headers.h"
 // clang-format on
 
-#include "iree/base/ref_ptr.h"
 #include "iree/base/status.h"
 #include "iree/base/synchronization.h"
 #include "iree/hal/vulkan/dynamic_symbols.h"
 #include "iree/hal/vulkan/extensibility_util.h"
 #include "iree/hal/vulkan/status_util.h"
+#include "iree/hal/vulkan/util/ref_ptr.h"
 
 namespace iree {
 namespace hal {
