+++ iree/hal/vulkan/util/ref_ptr_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iree/base/ref_ptr.h"
+#include "iree/hal/vulkan/util/ref_ptr.h"
 
 #include "iree/testing/gtest.h"
